# typescript/packages/apps/regions/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ARCADE Regions",
    "description": "Manages hierarchical structures of groups, regions, and zones.",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
      "def-0": {
        "description": "any non-searchable or filterable key:val attributes to specify metadata such as label, description etc.",
        "type": "object",
        "title": "attributes",
        "additionalProperties": {}
      },
      "def-1": {
        "description": "User-defined searchable tags",
        "type": "object",
        "title": "tags",
        "additionalProperties": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "null"
            }
          ]
        }
      },
...
      "def-8": {
        "type": "object",
        "properties": {
          "name": {
            "description": "The name of the Region.",
            "type": "string"
          },
          "attributes": {
            "$ref": "#/components/schemas/def-0"
          },
          "tags": {
            "$ref": "#/components/schemas/def-1"
          }
        },
        "title": "editRegionRequestBody"
      },
...
    }
  },
  "paths": {
...
    "/regions/{regionId}": {
      "get": {
        "operationId": "getRegion",
        "summary": "Retrieve a region.",
        "tags": [
          "Regions"
        ],
        "description": "Retrieve the details of a region.\n\nPermissions:\n- Only `reader` and above may view regions.\n",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "regionId",
            "required": true,
            "description": "Unique Region ID."
          },
          {
            "schema": {
              "default": "1.0.0",
              "type": "string"
            },
            "in": "header",
            "name": "accept-version",
            "required": true,
            "description": "API version"
          },
          {
            "schema": {
              "default": "application/json",
              "type": "string"
            },
            "in": "header",
            "name": "accept",
            "required": true,
            "description": "Accepted Content Type"
          }
        ],
        "x-security-scopes": [
          "admin",
          "contributor",
          "reader"
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Success.",
                  "type": "object",
                  "properties": {
                    "id": {
                      "description": "Unique Region ID.",
                      "type": "string"
                    },
                    "groupId": {
                      "description": "Unique Group ID.",
                      "type": "string"
                    },
                    "name": {
                      "description": "The name of the Region.",
                      "type": "string"
                    },
                    "attributes": {
                      "$ref": "#/components/schemas/def-0"
                    },
                    "tags": {
                      "$ref": "#/components/schemas/def-1"
                    },
                    "createdBy": {
                      "description": "ID of owner.",
                      "type": "string"
                    },
                    "createdAt": {
                      "description": "Date/time created",
                      "format": "date-time",
                      "type": "string"
                    },
                    "updatedBy": {
                      "description": "Last ID of user who made a change.",
                      "type": "string"
                    },
                    "updatedAt": {
                      "description": "Date/time updated",
                      "format": "date-time",
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "groupId",
                    "name",
                    "createdBy",
                    "createdAt"
                  ]
                },
                "examples": {
                  "Existing region": {
                    "summary": "Region retrieved successfully.",
                    "value": {
                      "id": "76ghytgt5",
                      "groupId": "sjduwydjd",
                      "name": "Abshire Harvest",
                      "createdAt": "2022-08-10T23:55:20.322Z",
                      "createdBy": "someone@somewhere.com"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Not found.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteRegion",
        "summary": "Deletes a region.",
        "tags": [
          "Regions"
        ],
        "description": "Deletes a region.\n\nPermissions:\n- Only `admin` and above may delete regions.\n",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "regionId",
            "required": true,
            "description": "Unique Region ID."
          },
          {
            "schema": {
              "default": "1.0.0",
              "type": "string"
            },
            "in": "header",
            "name": "accept-version",
            "required": true,
            "description": "API version"
          },
          {
            "schema": {
              "default": "application/json",
              "type": "string"
            },
            "in": "header",
            "name": "accept",
            "required": true,
            "description": "Accepted Content Type"
          }
        ],
        "x-security-scopes": [
          "admin"
        ],
        "responses": {
          "204": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Success.",
                  "type": "object",
                  "properties": {}
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Forbidden.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Not found.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateRegion",
        "summary": "Update a region.",
        "tags": [
          "Regions"
        ],
        "description": "Update an existing region.\n\nPermissions:\n- Only `admins` may update regions.\n",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/def-8"
              },
              "examples": {
                "Update region": {
                  "summary": "Update an existing region.",
                  "value": {
                    "name": "Abshire Harvest"
                  }
                },
                "Changing tags": {
                  "summary": "Changing the tags of a region.",
                  "value": {
                    "tags": {
                      "tier": "GOLD"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "regionId",
            "required": true,
            "description": "Unique Region ID."
          },
          {
            "schema": {
              "default": "1.0.0",
              "type": "string"
            },
            "in": "header",
            "name": "accept-version",
            "required": true,
            "description": "API version"
          },
          {
            "schema": {
              "default": "application/json",
              "type": "string"
            },
            "in": "header",
            "name": "accept",
            "required": true,
            "description": "Accepted Content Type"
          }
        ],
        "x-security-scopes": [
          "admin",
          "contributor"
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Success.",
                  "type": "object",
                  "properties": {
                    "id": {
                      "description": "Unique Region ID.",
                      "type": "string"
                    },
                    "groupId": {
                      "description": "Unique Group ID.",
                      "type": "string"
                    },
                    "name": {
                      "description": "The name of the Region.",
                      "type": "string"
                    },
                    "attributes": {
                      "$ref": "#/components/schemas/def-0"
                    },
                    "tags": {
                      "$ref": "#/components/schemas/def-1"
                    },
                    "createdBy": {
                      "description": "ID of owner.",
                      "type": "string"
                    },
                    "createdAt": {
                      "description": "Date/time created",
                      "format": "date-time",
                      "type": "string"
                    },
                    "updatedBy": {
                      "description": "Last ID of user who made a change.",
                      "type": "string"
                    },
                    "updatedAt": {
                      "description": "Date/time updated",
                      "format": "date-time",
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "groupId",
                    "name",
                    "createdBy",
                    "createdAt"
                  ]
                },
                "examples": {
                  "Existing region updated successfully": {
                    "summary": "Existing region created successfully.",
                    "value": {
                      "id": "0980yht42",
                      "groupId": "sjduwydjd",
                      "name": "Barrows Meadows",
                      "createdAt": "2022-08-10T23:55:20.322Z",
                      "createdBy": "someone@somewhere.com",
                      "updatedAt": "2022-08-12T07:34:23.132Z",
                      "updatedBy": "someoneelse@somewhere.com"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Bad request.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                },
                "examples": {
                  "Invalid request": {
                    "summary": "Invalid request.",
                    "value": {
                      "description": "Expected `formula` to be defined but not provided."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Forbidden.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Not found.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Conflict.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "syntaxErrors": {
                      "type": "object",
                      "properties": {
                        "charPositionInLine": {
                          "type": "integer"
                        },
                        "line": {
                          "type": "integer"
                        },
                        "msg": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "charPositionInLine",
                        "line",
                        "msg"
                      ]
                    }
                  },
                  "required": [
                    "message"
                  ]
                },
                "examples": {
                  "Name in use": {
                    "summary": "The `name` is already in use.",
                    "value": {
                      "description": "Name `xyz` already exists."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:30001"
    }
  ],
  "security": [],
...
}